requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
concurrent-log-handler==0.9.28
idna==3.11
paho-mqtt==2.1.0
portalocker==3.2.0
psycopg2-binary==2.9.10
PyYAML==6.0.3
requests==2.32.5
urllib3==2.5.0
